redesigned frontend, README updateed,login redirection

+++ frontend/src/app/loginform/loginform.component.ts
@@ -2,9 +2,9 @@ import { Component } from '@angular/core';
 
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { FormBuilder, Validators, FormGroup } from '@angular/forms';
-import { AuthService } from '../auth.service';
+import { AuthService } from '../service/auth.service';
 import { CookieService } from 'ngx-cookie-service';
-// import { Router } from '@angular/router';
+import { Router } from '@angular/router';
 // import { UserInterface } from '../user.interface';
 // import { tap } from 'rxjs';
 
@@ -14,7 +14,7 @@ import { CookieService } from 'ngx-cookie-service';
   styleUrl: './loginform.component.css',
 
   //standalone: true,
-  //imports: [MatIconModule, MatDividerModule, MatButtonModule],
+  //imports: [FormsModule, MatFormFieldModule, MatInputModule],
 })
 
 export class LoginformComponent {
@@ -23,18 +23,17 @@ export class LoginformComponent {
   loginForm: FormGroup;
 
   constructor(
-    private http: HttpClient, 
-    private authService: AuthService, 
-    private fb: FormBuilder, 
-    private cookieService: CookieService
+    private http: HttpClient,
+    private authService: AuthService,
+    private fb: FormBuilder,
+    private cookieService: CookieService,
+    private router: Router
   ) {
     this.loginForm = this.fb.group({
       username: ['', Validators.required],
       password: ['', Validators.required]
     });
   }
-
-  // router = inject(Router);
 
   onSubmit(): void {
     const formValues = this.loginForm.value;
@@ -57,33 +56,39 @@ export class LoginformComponent {
       )
       .subscribe((response) => {
         console.log(response.url);
-        if (response.url !== "http://localhost:5000/login/authFailed") {
-          alert('Authentication Success!')
-        }else{
-          alert('Authentication failed!')
+        if (response.status === 200 && response.url !== "http://localhost:5000/login/authFailed") {
+          //alert('Authentication Success!');
+          this.router.navigate(['contacts']);
+        } else {
+          this.router.navigate(['authFailed']);
         }
-      }, (error) => {alert('Authentication failed!')}
+      }, 
+      (error) => {
+        console.log(error)
+        //alert('Authentication failed!');
+        this.router.navigate(['authFailed']);
+      }
       );
 
   }
 
   getStatus(): void {
     this.http
-    .get<any>(
-      'http://localhost:5000/status/info',
-      {
-        observe: 'response', withCredentials: true
-      }
-    )
-    .subscribe((response) => {
-      let body = JSON.parse(JSON.stringify(response)).body;
-      console.log(body);
-      if (response.ok) {
-        alert('You are logged!')
-      } else {
-        alert('You are not authenticated!')
-      }
-    });
+      .get<any>(
+        'http://localhost:5000/status/info',
+        {
+          observe: 'response', withCredentials: true
+        }
+      )
+      .subscribe((response) => {
+        let body = JSON.parse(JSON.stringify(response)).body;
+        console.log(body);
+        if (response.ok) {
+          alert('You are logged!')
+        } else {
+          alert('You are not authenticated!')
+        }
+      });
 
   }
 
@@ -103,15 +108,17 @@ export class LoginformComponent {
           alert('Someting goes wrong!')
         }
       });
-      //Delete session from the cookies
-      this.deleteCookie('SESSION');
+    //Delete session from the cookies
+    //this.deleteCookie('SESSION');
   }
 
   deleteCookie(cookieName: string) {
     //Delete single cookie by name
-    this.cookieService.delete(cookieName);
-    //Delete all cookies
-    //this.cookieService.deleteAll();
+    if (this.cookieService.get(cookieName)) {
+      //Delete all cookies
+      //this.cookieService.deleteAll();
+      this.cookieService.delete(cookieName);
+    }
   }
 
   ngOnInit(): void {
